@JohanLindvall 2024-08-14
Hello, I just want to bring to your attention that this change makes our queries time out and/or consume a huge amount of CPU resources. See https://github.com/grafana/loki/issues/13869